requirements.txt
flask==3.1.0
flask-login==0.6.3
werkzeug>=3.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
openai==1.58.1
markdown2==2.4.12
